feat: 답글 삭제 시 검증하는 로직 추가 (#38)

+++ instragram/src/main/java/toyproject/instragram/reply/controller/ReplyApiController.java
@@ -5,6 +5,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.web.bind.annotation.*;
+import toyproject.instragram.common.auth.SignIn;
+import toyproject.instragram.common.auth.SignInMember;
 import toyproject.instragram.common.response.CommonSliceResponse;
 import toyproject.instragram.common.response.SliceInfo;
 import toyproject.instragram.reply.entity.Reply;
@@ -13,6 +15,9 @@ import toyproject.instragram.reply.service.ReplyService;
 
 import java.util.List;
 import java.util.stream.Collectors;
+
+import static toyproject.instragram.common.exception.api.ApiExceptionType.FORBIDDEN_COMMENT;
+import static toyproject.instragram.common.exception.api.ApiExceptionType.FORBIDDEN_REPLY;
 
 @RestController
 @RequiredArgsConstructor
@@ -25,15 +30,26 @@ public class ReplyApiController {
             @RequestParam Long commentId, @PageableDefault(size = 20) Pageable pageable) {
         Slice<Reply> replySlice = replyService.getReplySlice(commentId, pageable);
 
-        return new CommonSliceResponse<>(getReplyResponsesForm(replySlice.getContent()), SliceInfo.from(replySlice));
+        return new CommonSliceResponse<>(getReplyResponsesFrom(replySlice.getContent()), SliceInfo.from(replySlice));
     }
 
-    private List<ReplyResponse> getReplyResponsesForm(List<Reply> content) {
+    private List<ReplyResponse> getReplyResponsesFrom(List<Reply> content) {
         return content.stream().map(ReplyResponse::from).collect(Collectors.toList());
     }
 
     @DeleteMapping("/replies/{replyId}")
-    public void deleteReply(@PathVariable Long replyId) {
+    public void deleteReply(@SignIn SignInMember signInMember, @PathVariable Long replyId) {
+        validateReplyAccess(signInMember.getMemberId(), getReplyOwnerId(replyId));
         replyService.deleteReply(replyId);
+    }
+
+    private void validateReplyAccess(Long signInMemberId, Long replyOwnerId) {
+        if (!signInMemberId.equals(replyOwnerId)) {
+            throw FORBIDDEN_REPLY.getException();
+        }
+    }
+
+    private Long getReplyOwnerId(Long replyId) {
+        return replyService.getReply(replyId).getMember().getId();
     }
 }
